requirements.txt
django
django-smart-selects
django-bootstrap-v5
